feat(migration): alter "time_slot_minute" entity for SQLite
