+++ packages/contracts/src/index.ts
@@ -142,6 +142,7 @@ export * from './lib/tenant.model';
 export * from './lib/time-off.model';
 export * from './lib/timesheet-statistics.model';
 export * from './lib/timesheet.model';
+export * from './lib/time-slot-minute.model';
 export * from './lib/translation.model';
 export * from './lib/tree-node.model';
 export * from './lib/upwork.model';
