+++ packages/contracts/src/lib/timesheet.model.ts
@@ -1,4 +1,4 @@
-import { IBasePerTenantAndOrganizationEntityModel, IBaseRelationsEntityModel, ID, JsonData } from './base-entity.model';
+import { IBasePerTenantAndOrganizationEntityModel, IBaseRelationsEntityModel, ID } from './base-entity.model';
 import { IOrganizationContact, OrganizationContactBudgetTypeEnum } from './organization-contact.model';
 import {
 	IOrganizationProject,
@@ -14,6 +14,7 @@ import { IUser } from './user.model';
 import { IRelationalOrganizationTeam } from './organization-team.model';
 import { IScreenshot } from './screenshot.model';
 import { TimeFormatEnum } from './organization.model';
+import { ITimeSlotMinute } from './time-slot-minute.model';
 
 export interface ITimesheet extends IBasePerTenantAndOrganizationEntityModel {
 	employee: IEmployee;
@@ -228,42 +229,6 @@ export interface ITimeSlotTimeLogs extends IBasePerTenantAndOrganizationEntityMo
 	timeSlots: ITimeSlot[];
 	timeLogId: ID;
 	timeSlotId: ID;
-}
-
-export interface ITimeSlotMinute extends IBasePerTenantAndOrganizationEntityModel {
-	timeSlot?: ITimeSlot;
-	timeSlotId?: ID;
-
-	/**
-	 * Activity count metrics within the recorded minute.
-	 * - `keyboard`: Number of keyboard events
-	 * - `mouse`: Number of mouse events
-	 * - `location`: Number of movement events
-	 */
-	keyboard?: number;
-	mouse?: number;
-	location?: number;
-
-	// Activity timestamp
-	datetime?: Date;
-
-	/**
-	 * Detailed log of keyboard and mouse activity events for the minute,
-	 * such as click counts, key presses, or movement traces.
-	 */
-	kb_mouse_activity?: JsonData;
-
-	/**
-	 * Raw location-related activity data for the minute, such as GPS movement,
-	 * device orientation, or other physical motion events.
-	 */
-	location_activity?: JsonData;
-
-	/**
-	 * Custom activity data, allowing extensibility for domain-specific metrics
-	 * such as app usage, window focus, or custom plugins.
-	 */
-	custom_activity?: JsonData;
 }
 
 export interface IActivity extends IBasePerTenantAndOrganizationEntityModel {
