+++ packages/core/src/lib/core/entities/internal.ts
@@ -155,7 +155,7 @@ export * from '../../time-off-request/time-off-request.entity';
 export * from '../../time-tracking/activity/activity.entity';
 export * from '../../time-tracking/screenshot/screenshot.entity';
 export * from '../../time-tracking/time-log/time-log.entity';
-export * from '../../time-tracking/time-slot/time-slot-minute.entity';
+export * from '../../time-tracking/time-slot/time-slot-minute/time-slot-minute.entity';
 export * from '../../time-tracking/time-slot/time-slot.entity';
 export * from '../../time-tracking/timesheet/timesheet.entity';
 export * from '../../employee-notification/employee-notification.entity';
