+++ packages/core/src/lib/time-tracking/time-slot/commands/handlers/index.ts
@@ -1,7 +1,5 @@
 import { CreateTimeSlotHandler } from './create-time-slot.handler';
 import { UpdateTimeSlotHandler } from './update-time-slot.handler';
-import { CreateTimeSlotMinutesHandler } from './create-time-slot-minutes.handler';
-import { UpdateTimeSlotMinutesHandler } from './update-time-slot-minutes.handler';
 import { TimeSlotBulkCreateOrUpdateHandler } from './time-slot-bulk-create-or-update.handler';
 import { TimeSlotBulkCreateHandler } from './time-slot-bulk-create.handler';
 import { DeleteTimeSlotHandler } from './delete-time-slot.handler';
@@ -9,6 +7,8 @@ import { TimeSlotBulkDeleteHandler } from './time-slot-bulk-delete.handler';
 import { TimeSlotMergeHandler } from './time-slot-merge.handler';
 import { TimeSlotCreateHandler } from './time-slot-create.handler';
 import { ScheduleTimeSlotEntriesHandler } from './schedule-time-slot-entries.handler';
+import { UpdateTimeSlotMinutesHandler } from '../../time-slot-minute/commands/handlers/update-time-slot-minutes.handler';
+import { CreateTimeSlotMinutesHandler } from '../../time-slot-minute/commands/handlers/create-time-slot-minutes.handler';
 
 export const CommandHandlers = [
 	CreateTimeSlotHandler,
@@ -16,10 +16,10 @@ export const CommandHandlers = [
 	DeleteTimeSlotHandler,
 	TimeSlotBulkCreateOrUpdateHandler,
 	TimeSlotBulkCreateHandler,
-	CreateTimeSlotMinutesHandler,
-	UpdateTimeSlotMinutesHandler,
 	TimeSlotBulkDeleteHandler,
 	TimeSlotMergeHandler,
 	TimeSlotCreateHandler,
-	ScheduleTimeSlotEntriesHandler
+	ScheduleTimeSlotEntriesHandler,
+	CreateTimeSlotMinutesHandler,
+	UpdateTimeSlotMinutesHandler
 ];
