+++ packages/core/src/lib/time-tracking/time-slot/commands/index.ts
@@ -1,4 +1,3 @@
-export * from './create-time-slot-minutes.command';
 export * from './create-time-slot.command';
 export * from './delete-time-slot.command';
 export * from './time-slot-bulk-create-or-update.command';
@@ -6,6 +5,7 @@ export * from './time-slot-bulk-create.command';
 export * from './time-slot-create.command';
 export * from './time-slot-merge.command';
 export * from './time-slot-bulk-delete.command';
-export * from './update-time-slot-minutes.command';
 export * from './update-time-slot.command';
 export * from './schedule-time-slot-entries.command';
+export * from './../time-slot-minute/commands/create-time-slot-minutes.command';
+export * from './../time-slot-minute/commands/update-time-slot-minutes.command';
