+++ packages/core/src/lib/time-tracking/time-slot/time-slot-minute/commands/update-time-slot-minutes.command.ts
