+++ packages/core/src/lib/time-tracking/time-slot/time-slot-minute/repositories/mikro-orm-time-slot-minute.repository.ts
@@ -1,0 +1,4 @@
+import { MikroOrmBaseEntityRepository } from '../../../../core/repository/mikro-orm-base-entity.repository';
+import { TimeSlotMinute } from '../time-slot-minute.entity';
+
+export class MikroOrmTimeSlotMinuteRepository extends MikroOrmBaseEntityRepository<TimeSlotMinute> {}
