+++ packages/core/src/lib/time-tracking/time-slot/time-slot.module.ts
@@ -6,13 +6,13 @@ import { CommandHandlers } from './commands/handlers';
 import { RolePermissionModule } from '../../role-permission/role-permission.module';
 import { TimeSlot } from './time-slot.entity';
 import { TimeSlotController } from './time-slot.controller';
-import { TimeSlotMinute } from './time-slot-minute.entity';
 import { TimeSlotService } from './time-slot.service';
 import { TimeLogModule } from './../time-log/time-log.module';
 import { EmployeeModule } from './../../employee/employee.module';
 import { ActivityModule } from './../activity/activity.module';
 import { TypeOrmTimeSlotRepository } from './repository/type-orm-time-slot.repository';
-import { TypeOrmTimeSlotMinuteRepository } from './repository/type-orm-time-slot-minute.repository';
+import { TimeSlotMinute } from './time-slot-minute/time-slot-minute.entity';
+import { TypeOrmTimeSlotMinuteRepository } from './time-slot-minute/repositories/type-orm-time-slot-minute.repository';
 
 @Module({
 	controllers: [TimeSlotController],
